Ml2/requirements.txt
# Web Framework
Flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# ML Libraries
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2

# Model Persistence
joblib==1.3.2

# Database
mysql-connector-python==8.2.0

# Utilities
python-dotenv==1.0.0

# Visualization (for training notebooks)
matplotlib==3.8.2
seaborn==0.13.0

# Progress bars
tqdm==4.66.1
